:wrench: configurando arquico para nao mostrar o User admin

+++ jspprojeto-curso/src/main/java/dao/DAOUsuarioRepository.java
@@ -56,7 +56,7 @@ public class DAOUsuarioRepository {
 
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-		String sql = "SELECT * from model_login";
+		String sql = "SELECT * from model_login where useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
 
@@ -86,7 +86,7 @@ public class DAOUsuarioRepository {
 
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-		String sql = "SELECT * from model_login where upper(nome) like upper(?)";
+		String sql = "SELECT * from model_login where upper(nome) like upper(?) and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
 
@@ -113,7 +113,7 @@ public class DAOUsuarioRepository {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where upper(login) = upper('" + login + "')";
+		String sql = "select * from model_login where upper(login) = upper('" + login + "') and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
 
@@ -137,7 +137,7 @@ public class DAOUsuarioRepository {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where id = ?";
+		String sql = "select * from model_login where id = ? and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setLong(1, Long.parseLong(id));
@@ -170,7 +170,7 @@ public class DAOUsuarioRepository {
 	// metodo deletar user
 	public void deletarUser(String idUser) throws Exception {
 
-		String sql = "delete from model_login where id = ?";
+		String sql = "delete from model_login where id = ? and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
